+++ aldryn_events/cms_appconfig.py
@@ -7,13 +7,11 @@ from django.utils.translation import ugettext_lazy as _
 from django.utils.encoding import python_2_unicode_compatible
 
 from aldryn_apphooks_config.models import AppHookConfig
-from aldryn_reversion.core import version_controlled_content
 from cms.models.fields import PlaceholderField
 from parler.models import TranslatableModel, TranslatedFields
 
 
 @python_2_unicode_compatible
-@version_controlled_content
 class EventsConfig(TranslatableModel, AppHookConfig):
     """Adds some translatable, per-app-instance fields."""
     translations = TranslatedFields(