+++ aldryn_events/menu.py
@@ -51,4 +51,5 @@ class EventsMenu(CMSAttachMenu):
 
         return nodes
 
+
 menu_pool.register_menu(EventsMenu)