+++ aldryn_events/models.py
@@ -1,22 +1,8 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from distutils.version import LooseVersion
-
-try:
-    # Python>=2.7
-    from importlib import import_module
-except ImportError:
-    # Python==2.6
-    import warnings
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore", DeprecationWarning)
-        from django.utils.importlib import import_module
-
-from django import get_version
 from django.core.exceptions import ValidationError
 from django.core.urlresolvers import reverse
-from django.contrib.auth import get_user_model
 from django.db import models
 from django.utils import timezone
 from django.utils.encoding import force_text, python_2_unicode_compatible
@@ -29,13 +15,11 @@ from cms.utils.i18n import get_current_language, get_redirect_on_fallback
 from aldryn_translation_tools.models import (
     TranslationHelperMixin, TranslatedAutoSlugifyMixin,
 )
-from aldryn_reversion.core import version_controlled_content
 from djangocms_text_ckeditor.fields import HTMLField
 from extended_choices import Choices
 from filer.fields.image import FilerImageField
 from parler.models import TranslatableModel, TranslatedFields
 from sortedm2m.fields import SortedManyToManyField
-from reversion.revisions import default_revision_manager, RegistrationError
 
 from .cms_appconfig import EventsConfig
 from .conf import settings
@@ -43,72 +27,9 @@ from .managers import EventManager
 from .utils import get_additional_styles, date_or_datetime
 
 STANDARD = 'standard'
-strict_version = LooseVersion(get_version())
-
-# NOTE: We're using LooseVersion instead of StrictVersion because Aldryn
-# sometimes uses patched versions of Django in the form X.Y.Z.postN.
-if strict_version < LooseVersion('1.7.0'):
-    LTE_DJANGO_1_6 = True
-    # Prior to 1.7 it is pretty straight forward
-    user_model = get_user_model()
-    if user_model not in default_revision_manager.get_registered_models():
-        default_revision_manager.register(user_model)
-else:
-    # otherwise it is a pain, but thanks to solution of getting model from
-    # https://github.com/django-oscar/django-oscar/commit/c479a1983f326a9b059e157f85c32d06a35728dd
-    # we can do almost the same thing from the different side.
-    from django.apps import apps
-    from django.apps.config import MODELS_MODULE_NAME
-    from django.core.exceptions import AppRegistryNotReady
-
-    LTE_DJANGO_1_6 = False
-
-    def get_model(app_label, model_name):
-        """
-        Fetches a Django model using the app registry.
-        This doesn't require that an app with the given app label exists,
-        which makes it safe to call when the registry is being populated.
-        All other methods to access models might raise an exception about the
-        registry not being ready yet.
-        Raises LookupError if model isn't found.
-        """
-        try:
-            return apps.get_model(app_label, model_name)
-        except AppRegistryNotReady:
-            if apps.apps_ready and not apps.models_ready:
-                # If this function is called while `apps.populate()` is
-                # loading models, ensure that the module that defines the
-                # target model has been imported and try looking the model up
-                # in the app registry. This effectively emulates
-                # `from path.to.app.models import Model` where we use
-                # `Model = get_model('app', 'Model')` instead.
-                app_config = apps.get_app_config(app_label)
-                # `app_config.import_models()` cannot be used here because it
-                # would interfere with `apps.populate()`.
-                import_module('%s.%s' % (app_config.name, MODELS_MODULE_NAME))
-                # In order to account for case-insensitivity of model_name,
-                # look up the model through a private API of the app registry.
-                return apps.get_registered_model(app_label, model_name)
-            else:
-                # This must be a different case (e.g. the model really doesn't
-                # exist). We just re-raise the exception.
-                raise
-
-    # now get the real user model
-    user_model = getattr(settings, 'AUTH_USER_MODEL', 'auth.User')
-    model_app_name, model_model = user_model.split('.')
-    user_model_object = get_model(model_app_name, model_model)
-    # and try to register, if we have a registration error - that means that
-    # it has been registered already
-
-    try:
-        default_revision_manager.register(user_model_object)
-    except RegistrationError:
-        pass
 
 
 @python_2_unicode_compatible
-@version_controlled_content(follow=['event_coordinators', 'app_config'])
 class Event(TranslatedAutoSlugifyMixin,
             TranslationHelperMixin,
             TranslatableModel):
@@ -309,7 +230,6 @@ class Event(TranslatedAutoSlugifyMixin,
 
 
 @python_2_unicode_compatible
-@version_controlled_content(follow=['user'])
 class EventCoordinator(models.Model):
 
     name = models.CharField(max_length=200, blank=True)
@@ -401,19 +321,11 @@ class BaseEventPlugin(CMSPlugin):
     # with any other plugins that have a field with the same name as the
     # lowercase of the class name of this model.
     # https://github.com/divio/django-cms/issues/5030
-    if LTE_DJANGO_1_6:
-        # related_name='%(app_label)s_%(class)s' does not work on  Django 1.6
-        cmsplugin_ptr = models.OneToOneField(
-            CMSPlugin,
-            related_name='+',
-            parent_link=True,
-        )
-    else:
-        cmsplugin_ptr = models.OneToOneField(
-            CMSPlugin,
-            related_name='%(app_label)s_%(class)s',
-            parent_link=True,
-        )
+    cmsplugin_ptr = models.OneToOneField(
+        CMSPlugin,
+        related_name='%(app_label)s_%(class)s',
+        parent_link=True,
+    )
 
     def copy_relations(self, old_instance):
         self.app_config = old_instance.app_config